Added bower.json and package.json with both declaring terraformer and terraformer-arcgis-parser as dependencies.
