+++ .gitignore
@@ -1,8 +1,6 @@
 ### JetBrains template
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio
 
-*.iml
-
 ## Directory-based project format:
 .idea/
 # if you remove the above rule, at least ignore the following:
@@ -104,30 +102,35 @@ sftp-config.json
 *.un~
 Session.vim
 .netrwhist
-*~
+
 ### TortoiseGit template
 # Project-level settings
 /.tgitconfig
+
 ### TextMate template
 *.tmproj
 *.tmproject
 tmtags
+
 ### SVN template
 .svn/
+
 ### Vagrant template
 .vagrant/
+
 ### NetBeans template
 nbproject/private/
-build/
 nbbuild/
 dist/
 nbdist/
 nbactions.xml
 nb-configuration.xml
 .nb-gradle/
+
 ### LibreOffice template
 # LibreOffice locks
 .~lock.*#
+
 ### Eclipse template
 *.pydevproject
 .metadata
@@ -135,7 +138,6 @@ nb-configuration.xml
 bin/
 tmp/
 *.tmp
-*.bak
 *.swp
 *~.nib
 local.properties
@@ -168,13 +170,14 @@ local.properties
 
 # TeXlipse plugin
 .texlipse
+
 ### Dreamweaver template
 # DW Dreamweaver added files
 _notes
 dwsync.xml
+
 ### Emacs template
 # -*- mode: gitignore; -*-
-*~
 \#*\#
 /.emacs.desktop
 /.emacs.desktop.lock
@@ -205,10 +208,12 @@ tramp
 
 # cask packages
 .cask/
+
 ### CodeKit template
 # General CodeKit files to ignore
 config.codekit
 /min
+
 ### OSX template
 .DS_Store
 .AppleDouble
@@ -234,6 +239,7 @@ Icon
 Network Trash Folder
 Temporary Items
 .apdisk
+
 ### Windows template
 # Windows image file caches
 Thumbs.db
@@ -257,4 +263,7 @@ $RECYCLE.BIN/
 /.hg/*
 */.hg/*
 .hgignore
+
+node_modules
+bower_components
